requirements.txt
Flask
Flask-Cors
gunicorn
requests
firebase-admin
Werkzeug
python-dotenv
